pyproject.toml
[project]
name = "mawa"
version = "0.1.0"
description = ""
authors = [
    {name = "Tomas Jelinek",email = "jelkosz@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.1,<0.35.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "google-genai (>=1.10.0,<2.0.0)",
    "google-adk (>=0.3.0,<0.4.0)",
    "cachetools (>=5.5.2,<6.0.0)",
    "diskcache (>=5.6.3,<6.0.0)",
    "mcp[cli] (>=1.11.0,<2.0.0)",
    "deprecated (>=1.2.18,<2.0.0)",
]

[tool.poetry]
packages = [{include = "mawa", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
